MNT: sensors to sensor rename imports
