+++ rocketpy/sensors/__init__.py
@@ -1,4 +1,4 @@
 from .accelerometer import Accelerometer
 from .barometer import Barometer
 from .gyroscope import Gyroscope
-from .sensors import InertialSensor, ScalarSensor, Sensor
+from .sensor import InertialSensor, ScalarSensor, Sensor
